requirements.txt
discord-py-interactions==4.3.4
psycopg2==2.9.5